% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_olsLM.R
\name{plot.olsLM}
\alias{plot.olsLM}
\title{Diagnostic plots for olsLM objects}
\usage{
\method{plot}{olsLM}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"olsLM"}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Produces four standard diagnostic plots for linear regression,
similar to \code{plot(lm)}:
\enumerate{
\item Residuals vs Fitted
\item Normal Q-Q Plot
\item Scale-Location Plot
\item Residuals vs Leverage
}
}
