% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_olsLM.R
\name{summary.olsLM}
\alias{summary.olsLM}
\title{Summary information for olsLM()}
\usage{
\method{summary}{olsLM}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"olsLM"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
The input object, invisibly.
}
\description{
This function prints a summary of an \code{olsLM} fitted model,
including coefficient table, residual standard error, and
R-squared statistics.
}
